Move utils to common
